Move cli module out of duffy.app
We'll have subcommands soon, and most of them aren't about running the
web app.
