+++ pyproject.toml
@@ -42,4 +42,4 @@ requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-duffy = "duffy.app.cli:main"
+duffy = "duffy.cli:main"
