+++ tests/test_cli.py
@@ -4,11 +4,11 @@ from unittest import mock
 import pytest
 from click.testing import CliRunner
 
-from duffy.app.cli import main
+from duffy.cli import main
 from duffy.version import __version__
 
 HERE = Path(__file__).parent
-EXAMPLE_CONFIG = HERE.parent.parent / "etc" / "duffy-example-config.yaml"
+EXAMPLE_CONFIG = HERE.parent / "etc" / "duffy-example-config.yaml"
 
 
 def test_cli_version():
@@ -36,7 +36,7 @@ def test_cli_suggestion():
 @pytest.mark.parametrize(
     "parameters", ((), ("--host=127.0.0.1",), (f"--config={EXAMPLE_CONFIG.absolute()}",))
 )
-@mock.patch("duffy.app.cli.uvicorn.run")
+@mock.patch("duffy.cli.uvicorn.run")
 def test_cli_main(uvicorn_run, parameters):
     runner = CliRunner()
     runner.invoke(main, parameters)
